Refactored precall native
